Phase A/B: split OGC managers and controllers, add multispectral gating, OgcAuth + WmsValidator + tests
